[Test] Ensourced accesspath-base.
Moved the test next to the code it calls.
