+++ lib/SILOptimizer/UtilityPasses/TestRunner.cpp
@@ -311,43 +311,6 @@ static FunctionTest FieldSensitiveMultiDefUseLiveRangeTest(
       boundary.print(llvm::errs());
     });
 
-//===----------------------------------------------------------------------===//
-// MARK: AccessPath Unit Tests
-//===----------------------------------------------------------------------===//
-
-struct AccessUseTestVisitor : public AccessUseVisitor {
-  AccessUseTestVisitor()
-    : AccessUseVisitor(AccessUseType::Overlapping,
-                       NestedAccessType::IgnoreAccessBegin) {}
-
-  bool visitUse(Operand *op, AccessUseType useTy) override {
-    switch (useTy) {
-    case AccessUseType::Exact:
-      llvm::errs() << "Exact Use: ";
-      break;
-    case AccessUseType::Inner:
-      llvm::errs() << "Inner Use: ";
-      break;
-    case AccessUseType::Overlapping:
-      llvm::errs() << "Overlapping Use ";
-      break;
-    }
-    llvm::errs() << *op->getUser();
-    return true;
-  }
-};
-
-static FunctionTest AccessPathBaseTest("accesspath-base", [](auto &function,
-                                                             auto &arguments,
-                                                             auto &test) {
-  auto value = arguments.takeValue();
-  function.dump();
-  llvm::outs() << "Access path base: " << value;
-  auto accessPathWithBase = AccessPathWithBase::compute(value);
-  AccessUseTestVisitor visitor;
-  visitAccessPathBaseUses(visitor, accessPathWithBase, &function);
-});
-
 } // namespace swift::test
 
 //===----------------------------------------------------------------------===//
